Show appeal ruling modified history for indictment cases
