+++ apps/judicial-system/web/pages/verjandi/krafa/[id].ts
@@ -1,3 +1,3 @@
-import { CaseOverview } from '@island.is/judicial-system-web/src/routes/Defender/RestrictionCase/CaseOverview'
+import { CaseOverview } from '@island.is/judicial-system-web/src/routes/Defender/RequestCase/CaseOverview'
 
 export default CaseOverview
