+++ apps/judicial-system/web/src/components/index.ts
@@ -138,6 +138,8 @@ export { VictimInfo } from './VictimInfo/VictimInfo'
 export { LabelValue } from './LabelValue/LabelValue'
 export { default as IconButton } from './IconButton/IconButton'
 export { LawyerRegistryContext } from './LawyerRegistryProvider/LawyerRegistryProvider'
-export { default as ArraignmentAlert } from './ArraignmentAlert/ArraignmentAlert'
+export { default as ArraignmentAlert } from './Alerts/ArraignmentAlert/ArraignmentAlert'
+export { default as RulingModifiedAlert } from './Alerts/RulingModifiedAlert/RulingModifiedAlert'
+export { default as AppealRulingModifiedAlert } from './Alerts/AppealRulingModifiedAlert/AppealRulingModifiedAlert'
 export { default as ChangeProsecutorModal } from './Modals/ChangeProsecutorModal/ChangeProsecutorModal'
 export { default as TinyMCE } from './TinyMCE/TinyMCE'
