+++ apps/judicial-system/web/src/routes/CourtOfAppeal/Result/Result.strings.tsx
@@ -7,12 +7,6 @@ export const result = defineMessages({
     description:
       'Notaður sem texti í "Næsta" hnapp í úrskurðarferli héraðsdóms',
   },
-  rulingModifiedTitle: {
-    id: 'judicial.system.core:court_of_appeal.result.ruling_modified_title',
-    defaultMessage: 'Úrskurður leiðréttur',
-    description:
-      'Notaður sem titill í "Úrskurður leiðréttur" hlutanum í úrskurðarferli héraðsdóms',
-  },
   requestAppealRulingNotToBePublished: {
     id: 'judicial.system.core:result.request_appeal_ruling_not_to_be_published',
     defaultMessage:
