+++ apps/judicial-system/web/src/routes/Defender/IndictmentCase/IndictmentOverview.tsx
@@ -15,19 +15,19 @@ import { titles } from '@island.is/judicial-system-web/messages'
 import {
   AllIndictmentCaseFiles,
   AlternativeServiceAnnouncement,
+  AppealRulingModifiedAlert,
   Conclusion,
   CourtCaseInfo,
   FormContentContainer,
   FormContext,
   FormFooter,
   IndictmentCaseScheduledCard,
-  // IndictmentsLawsBrokenAccordionItem, NOTE: Temporarily hidden while list of laws broken is not complete
   InfoCardActiveIndictment,
   InfoCardClosedIndictment,
-  MarkdownWrapper,
   PageHeader,
   PageLayout,
   PageTitle,
+  RulingModifiedAlert,
   serviceAnnouncementsStrings,
   UserContext,
   ZipButton,
@@ -155,20 +155,6 @@ const IndictmentOverview: FC = () => {
         <FormContentContainer>
           <PageTitle>{caseIsClosed ? 'Máli lokið' : 'Yfirlit ákæru'}</PageTitle>
           <CourtCaseInfo workingCase={workingCase} />
-          {workingCase.rulingModifiedHistory && (
-            <Box marginBottom={5}>
-              <AlertMessage
-                type="info"
-                title="Mál leiðrétt"
-                message={
-                  <MarkdownWrapper
-                    markdown={workingCase.rulingModifiedHistory}
-                    textProps={{ variant: 'small' }}
-                  />
-                }
-              />
-            </Box>
-          )}
           {workingCase.reopenReason && !isCompletedCase(workingCase.state) && (
             <Box marginBottom={2}>
               <AlertMessage
@@ -221,6 +207,8 @@ const IndictmentOverview: FC = () => {
             </Fragment>
           ))}
           <div className={grid({ gap: 5, marginBottom: 10 })}>
+            <AppealRulingModifiedAlert />
+            <RulingModifiedAlert />
             {caseHasBeenReceivedByCourt &&
               workingCase.court &&
               latestDate?.date &&
