--- apps/judicial-system/web/src/routes/Shared/SignedVerdictOverview/SignedVerdictOverview.strings.ts
@@ -1,10 +1,0 @@
-import { defineMessages } from 'react-intl'
-
-export const strings = defineMessages({
-  rulingModifiedTitle: {
-    id: 'judicial.system.core:signed_verdict_overview.ruling_modified_title',
-    defaultMessage: 'Úrskurður leiðréttur',
-    description:
-      'Notaður sem titill í leiðréttingu úrskurðar landsréttar í yfirliti lokins máls.',
-  },
-})
